+++ pages/roadmaps/new.tsx
@@ -8,7 +8,7 @@ import type { Library, Tag } from '$prisma/client'
 import { RoadmapCreateBody } from '~/server/types'
 import { createReqSteps, createReqTags, createRoadmap } from '~/utils/roadmaps'
 import Layout from '~/components/Layout'
-import { HEADER_BTN_TYPES } from '~/components/header/Header'
+import { HEADER_BTN_TYPES } from '~/components/Header'
 import { SubmitHandler } from 'react-hook-form'
 import { SelectOption } from '~/components/parts/SelectInput'
 import { useRouter } from 'next/router'
